Gemini/4.Text_to_SQL_LLM_app/requirements.txt
streamlit 
ipykernel 
python-dotenv 
google-generativeai
pymysql
